🐛 Averts Failed Fitz Usages for Special Paths
When PDF paths contain special characters (that are otherwise legal), Fitz can fail to open the PDF. This averts that by using temporary files with safe names for the PDFs.

The DOCX handling was _also_ updated to use temporary file handling so that it would remain consistent with PDF handling.