🚸 Adds PDF/DOCX Export Retries w/ Exponential Backoff
Sometimes OneNote gets ornery and we need to wait a bit and retry.

In truth, this turned out to not avert the specific problem I'm encountering, (whose nature is a OneNote page that refuses to export as DOCX or MHT). However,  it is advantageous to have this level of resilience in place, so we are keeping this for the time being.