🏗️ Replaces DOCX-based Export w/ MHTML-based Export
Previously we had been getting content out of OneNote by instructing OneNote to export DOCX files for us. What wasn't obvious was that OneNote, in fact, implements such behavior by temporarily exporting to MHTML and then converting the MHTML to a DOCX before returning to the caller. This switches us to directly instruct OneNote to export the MHTML itself, which we then read and extract ourselves.

This vastly speeds up the acquisition of a Pandoc AST for the export pipeline, while also improving the fidelity of the exported content itself.

The pre-existing DOCX-based export behavior is still available behind a flag, but its maintenance will not be a priority (and can expected to be removed without notice).